Revert "Solution. Refactor #1"
This reverts commit ae1e8e52ae2502c70c36be96b70b378e9e28450d.

+++ ApiTemplate.Application/Services/AccountService.cs
@@ -32,9 +32,7 @@ namespace ApiTemplate.Application.Services
 
         public async Task<AccountDTO> GetAccount(int userId)
         {
-            var appUser = await _signInManager.UserManager.FindByIdAsync(userId.ToString())
-                ?? throw new MyApplicationException(ErrorStatus.NotFound, "User not found");
-
+            var appUser = await _signInManager.UserManager.FindByIdAsync(userId.ToString()) ?? throw new MyApplicationException(ErrorStatus.NotFound, "User not found");
             return _mapper.Map<AccountDTO>(appUser);
         }
 
@@ -45,8 +43,7 @@ namespace ApiTemplate.Application.Services
             var toInsert = _mapper.Map<AccountEntity>(model);
 
             var res = await _signInManager.UserManager.CreateAsync(toInsert, model.Password);
-            if (!res.Succeeded)
-                throw new MyApplicationException(ErrorStatus.InvalidData, string.Join(" ", res.Errors.Select(c => c.Description)));
+            if (!res.Succeeded) throw new MyApplicationException(ErrorStatus.InvalidData, string.Join(" ", res.Errors.Select(c => c.Description)));
 
             await _signInManager.UserManager.AddToRoleAsync(toInsert, model.Role.ToString());
         }
@@ -54,10 +51,9 @@ namespace ApiTemplate.Application.Services
         private async Task<AccountEntity> DeleteSameNotConfirmed(string email)
         {
             var existingUser = await _signInManager.UserManager.FindByEmailAsync(email);
-            if (existingUser == null)
-                return null;
+            if (existingUser == null) return null;
 
-            if (await existingUser.IsEmailConfirmedAsync(_signInManager.UserManager))
+            if (existingUser.EmailConfirmed)
                 throw new MyApplicationException(ErrorStatus.InvalidData, "User already exists");
 
             await _signInManager.UserManager.DeleteAsync(existingUser);
@@ -66,15 +62,11 @@ namespace ApiTemplate.Application.Services
 
         public async Task<RefreshTokenDTO> LoginAccount(LoginAccountDTO model)
         {
-            var appUser = await _signInManager.UserManager.FindByEmailAsync(model.Email)
-                ?? throw new MyApplicationException(ErrorStatus.NotFound, "User not found");
-
-            if (!await appUser.IsEmailConfirmedAsync(_signInManager.UserManager))
-                throw new MyApplicationException(ErrorStatus.InvalidData, "Email unconfirmed");
+            var appUser = await _signInManager.UserManager.FindByEmailAsync(model.Email) ?? throw new MyApplicationException(ErrorStatus.NotFound, "User not found");
+            if (!await _signInManager.UserManager.IsEmailConfirmedAsync(appUser)) throw new MyApplicationException(ErrorStatus.InvalidData, "Email unconfirmed");
 
             var res = await _signInManager.PasswordSignInAsync(appUser, model.Password, model.RememberMe, false);
-            if (!res.Succeeded)
-                throw new MyApplicationException(ErrorStatus.InvalidData, "Password or user invalid");
+            if (!res.Succeeded) throw new MyApplicationException(ErrorStatus.InvalidData, "Password or user invalid");
 
             return await CreateNewJwtPair(appUser);
         }
@@ -83,7 +75,8 @@ namespace ApiTemplate.Application.Services
         {
             var appUser = await _signInManager.UserManager.FindByIdAsync(userId.ToString());
 
-            AccountEntity.ValidateRefreshToken(appUser, model.RefreshToken);
+            if (appUser == null || appUser.RefreshToken != model.RefreshToken || appUser.RefreshTokenExpiryTime <= DateTime.UtcNow)
+                throw new MyApplicationException(ErrorStatus.Unauthorized);
 
             return await CreateNewJwtPair(appUser);
         }
@@ -99,9 +92,7 @@ namespace ApiTemplate.Application.Services
 
         public async Task SendDigitCodeByEmail(string email)
         {
-            var appUser = await _signInManager.UserManager.FindByEmailAsync(email)
-                ?? throw new MyApplicationException(ErrorStatus.NotFound, "User not found");
-
+            var appUser = await _signInManager.UserManager.FindByEmailAsync(email) ?? throw new MyApplicationException(ErrorStatus.NotFound, "User not found");
             var asString = TokenEnum.EmailToken.ToString();
 
             var userEmailTokens = await _userTokenRepo.GetIQueryable()
@@ -112,9 +103,13 @@ namespace ApiTemplate.Application.Services
 
             var digitCode = new Random().Next(1111, 9999).ToString("D4");
 
-            var token = await AccountTokenEntity.Create(appUser, digitCode, asString, _signInManager.UserManager);
-
-            await _userTokenRepo.InsertAsync(token, true);
+            await _userTokenRepo.InsertAsync(new AccountTokenEntity
+            {
+                UserId = appUser.Id,
+                Name = digitCode,
+                LoginProvider = asString,
+                Value = await _signInManager.UserManager.GenerateEmailConfirmationTokenAsync(appUser)
+            }, true);
 
             await _emailTemplateService.SendDigitCodeAsync(new EmailDTO
             {
@@ -132,11 +127,13 @@ namespace ApiTemplate.Application.Services
                                              .Where(x => x.Name == digitCode && x.LoginProvider == TokenEnum.EmailToken.ToString())
                                              .ToListAsync();
 
-            AccountTokenEntity.EnsureNoDuplicateTokensByNameAndProvider(tokens);
+            if (tokens.Count > 1 || tokens.Count == 0)
+                throw new MyApplicationException(ErrorStatus.InvalidData, "This code is invalid, please request a new one");
 
             var token = tokens.First();
-
-            await token.User.ConfirmEmailAsync(_signInManager.UserManager, token);
+            var result = await _signInManager.UserManager.ConfirmEmailAsync(token.User, token.Value);
+            if (!result.Succeeded)
+                throw new MyApplicationException(ErrorStatus.InvalidData, result.Errors.FirstOrDefault()!.Description);
 
             await _userTokenRepo.DeleteAsync(token, true);
         }
